fixed readme file

+++ auto-install/conf/install-java8.sh
@@ -3,6 +3,7 @@
 echo "Loading properties"
 . install.properties
 
+if [ ! -e "/opt/jdk${JAVA_VERSION}" ]; then
 tar zxvf /root/Downloads/${JDK_PACKAGE}.tar.gz
-
 mv jdk${JAVA_VERSION} /opt
+fi
